UniStore: Enable DataSyncer Mode1 + better logging
