+++ pkg/apiserver/rest/dualwriter_mode2.go
@@ -2,21 +2,16 @@ package rest
 
 import (
 	"context"
-	"fmt"
 	"time"
 
-	"github.com/prometheus/client_golang/prometheus"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/api/meta"
 	metainternalversion "k8s.io/apimachinery/pkg/apis/meta/internalversion"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apiserver/pkg/endpoints/request"
 	"k8s.io/apiserver/pkg/registry/rest"
 	"k8s.io/klog/v2"
 
-	"github.com/grafana/authlib/claims"
-	"github.com/grafana/grafana/pkg/apimachinery/identity"
 	"github.com/grafana/grafana/pkg/apimachinery/utils"
 )
 
@@ -391,214 +386,4 @@ func enrichLegacyObject(originalObj, returnedObj runtime.Object) error {
 	accessorReturned.SetResourceVersion(accessorOriginal.GetResourceVersion())
 	accessorReturned.SetUID(accessorOriginal.GetUID())
 	return nil
-}
-
-func getSyncRequester(orgId int64) *identity.StaticRequester {
-	return &identity.StaticRequester{
-		Type:           claims.TypeServiceAccount, // system:apiserver
-		UserID:         1,
-		OrgID:          orgId,
-		Name:           "admin",
-		Login:          "admin",
-		OrgRole:        identity.RoleAdmin,
-		IsGrafanaAdmin: true,
-		Permissions: map[int64]map[string][]string{
-			orgId: {
-				"*": {"*"}, // all resources, all scopes
-			},
-		},
-	}
-}
-
-type syncItem struct {
-	name       string
-	objStorage runtime.Object
-	objLegacy  runtime.Object
-}
-
-func getList(ctx context.Context, obj rest.Lister, listOptions *metainternalversion.ListOptions) ([]runtime.Object, error) {
-	ll, err := obj.List(ctx, listOptions)
-	if err != nil {
-		return nil, err
-	}
-
-	return meta.ExtractList(ll)
-}
-
-func mode2DataSyncer(ctx context.Context, legacy LegacyStorage, storage Storage, resource string, reg prometheus.Registerer, serverLockService ServerLockService, requestInfo *request.RequestInfo) (bool, error) {
-	metrics := &dualWriterMetrics{}
-	metrics.init(reg)
-
-	log := klog.NewKlogr().WithName("DualWriterMode2Syncer")
-
-	everythingSynced := false
-	outOfSync := 0
-	syncSuccess := 0
-	syncErr := 0
-
-	maxInterval := dataSyncerInterval + 5*time.Minute
-
-	var errSync error
-	const maxRecordsSync = 1000
-
-	// LockExecuteAndRelease ensures that just a single Grafana server acquires a lock at a time
-	// The parameter 'maxInterval' is a timeout safeguard, if the LastExecution in the
-	// database is older than maxInterval, we will assume the lock as timeouted. The 'maxInterval' parameter should be so long
-	// that is impossible for 2 processes to run at the same time.
-	err := serverLockService.LockExecuteAndRelease(ctx, "dualwriter mode 2 sync", maxInterval, func(context.Context) {
-		log.Info("starting dualwriter mode 2 sync")
-		startSync := time.Now()
-
-		orgId := int64(1)
-
-		ctx = klog.NewContext(ctx, log)
-		ctx = identity.WithRequester(ctx, getSyncRequester(orgId))
-		ctx = request.WithNamespace(ctx, requestInfo.Namespace)
-		ctx = request.WithRequestInfo(ctx, requestInfo)
-
-		storageList, err := getList(ctx, storage, &metainternalversion.ListOptions{
-			Limit: maxRecordsSync,
-		})
-		if err != nil {
-			log.Error(err, "unable to extract list from storage")
-			return
-		}
-
-		if len(storageList) >= maxRecordsSync {
-			errSync = fmt.Errorf("unified storage has more than %d records. Aborting sync", maxRecordsSync)
-			log.Error(errSync, "Unified storage has more records to be synced than allowed")
-			return
-		}
-
-		log.Info("got items from unified storage", "items", len(storageList))
-
-		legacyList, err := getList(ctx, legacy, &metainternalversion.ListOptions{})
-		if err != nil {
-			log.Error(err, "unable to extract list from legacy storage")
-			return
-		}
-		log.Info("got items from legacy storage", "items", len(legacyList))
-
-		itemsByName := map[string]syncItem{}
-		for _, obj := range legacyList {
-			accessor, err := utils.MetaAccessor(obj)
-			if err != nil {
-				log.Error(err, "error retrieving accessor data for object from legacy storage")
-				continue
-			}
-			name := accessor.GetName()
-
-			item, ok := itemsByName[name]
-			if !ok {
-				item = syncItem{}
-			}
-			item.name = name
-			item.objLegacy = obj
-			itemsByName[name] = item
-		}
-
-		for _, obj := range storageList {
-			accessor, err := utils.MetaAccessor(obj)
-			if err != nil {
-				log.Error(err, "error retrieving accessor data for object from storage")
-				continue
-			}
-			name := accessor.GetName()
-
-			item, ok := itemsByName[name]
-			if !ok {
-				item = syncItem{}
-			}
-			item.name = name
-			item.objStorage = obj
-			itemsByName[name] = item
-		}
-		log.Info("got list of items to be synced", "items", len(itemsByName))
-
-		for name, item := range itemsByName {
-			// upsert if:
-			// - existing in both legacy and storage, but objects are different, or
-			// - if it's missing from storage
-			if item.objLegacy != nil &&
-				((item.objStorage != nil && !Compare(item.objLegacy, item.objStorage)) || (item.objStorage == nil)) {
-				outOfSync++
-
-				accessor, err := utils.MetaAccessor(item.objLegacy)
-				if err != nil {
-					log.Error(err, "error retrieving accessor data for object from storage")
-					continue
-				}
-
-				if item.objStorage != nil {
-					accessorStorage, err := utils.MetaAccessor(item.objStorage)
-					if err != nil {
-						log.Error(err, "error retrieving accessor data for object from storage")
-						continue
-					}
-					accessor.SetResourceVersion(accessorStorage.GetResourceVersion())
-					accessor.SetUID(accessorStorage.GetUID())
-
-					log.Info("updating item on unified storage", "name", name)
-				} else {
-					accessor.SetResourceVersion("")
-					accessor.SetUID("")
-
-					log.Info("inserting item on unified storage", "name", name)
-				}
-
-				objInfo := rest.DefaultUpdatedObjectInfo(item.objLegacy, []rest.TransformFunc{}...)
-				res, _, err := storage.Update(ctx,
-					name,
-					objInfo,
-					func(ctx context.Context, obj runtime.Object) error { return nil },
-					func(ctx context.Context, obj, old runtime.Object) error { return nil },
-					true, // force creation
-					&metav1.UpdateOptions{},
-				)
-				if err != nil {
-					log.WithValues("object", res).Error(err, "could not update in storage")
-					syncErr++
-				} else {
-					syncSuccess++
-				}
-			}
-
-			// delete if object does not exists on legacy but exists on storage
-			if item.objLegacy == nil && item.objStorage != nil {
-				outOfSync++
-
-				ctx = request.WithRequestInfo(ctx, &request.RequestInfo{
-					APIGroup:  requestInfo.APIGroup,
-					Resource:  requestInfo.Resource,
-					Name:      name,
-					Namespace: requestInfo.Namespace,
-				})
-
-				log.Info("deleting item from unified storage", "name", name)
-
-				deletedS, _, err := storage.Delete(ctx, name, func(ctx context.Context, obj runtime.Object) error { return nil }, &metav1.DeleteOptions{})
-				if err != nil {
-					if !apierrors.IsNotFound(err) {
-						log.WithValues("objectList", deletedS).Error(err, "could not delete from storage")
-					}
-					syncErr++
-				} else {
-					syncSuccess++
-				}
-			}
-		}
-
-		everythingSynced = outOfSync == syncSuccess
-
-		metrics.recordDataSyncerOutcome(mode2Str, resource, everythingSynced)
-		metrics.recordDataSyncerDuration(err != nil, mode2Str, resource, startSync)
-
-		log.Info("finished syncing items", "items", len(itemsByName), "updated", syncSuccess, "failed", syncErr, "outcome", everythingSynced)
-	})
-
-	if errSync != nil {
-		err = errSync
-	}
-
-	return everythingSynced, err
 }
